Write real times in tar entries
Of course this means that you're pretty much guaranteed a different OCI
archive for multiple builds with the same binary, i.e. you don't get the
kind of reproducibility that something like ko provides. However this is
more similar to how other builders work, and it's not like our archives
were identical in the first place since we write the current time into
our manifests.