Switch to go-containerregistry for the build
This should fully get rid of the use of my custom ocibuild package, and
get us more on the way to having built-in registry support.

Originally I thought about implementing v1.Image based on OCI archives,
but I decided to just go with the Docker archive support provided by
go-containerregistry out of the box. So we now produce Docker archives
instead of OCI archives, which requires that we include a tag. I just
hacked one in place for now.