Switch CLI to cobra; 'zeroimage' to 'zeroimage build'
Looking toward a future desire to integrate container registry support
into zeroimage, I want to open up the possibility of things like
'zeroimage login' and a more comprehensive set of build flags.

Yes, I removed the "few dependencies outside of the Go standard library"
marketing. I NEVER should have written that line in the first place, as
it's quite a reductive view of the situation. The truth is that there's
a massive dependency outside of the Go standard library, in the need to
have an external tool like Skopeo to handle all interaction with OCI
registries and container runtimes, i.e. the things that are useful to
build an image for, or to get base images from. I think that zeroimage
should better encompass that kind of complexity in order to simplify the
ultimate user experience.