Enable deferred blob loading in ocibuild
It's no longer a hard requirement to have the blob fully buffered into
memory, as long as the descriptor and diff ID are known in advance. This
should help with base images from registries, so that we don't always
have to download their content when the relevant blobs already exist in
the target registry.