Split up the ocibuild package
This change splits up ocibuild into three separate packages.

The image package is a simple "types" package that provides a common
interface to represent container images based on OCI types.

The ociarchive package deals with OCI layout tar archives, both for
reading and writing.

The tarlayer package automatically compresses and digests the contents
of a tar archive built with the tarbuild package.

This should make it easier to separate the interactions between archive
related things and registry related things.