Add basic support for Docker manifest types
This will be important to pull from registries. The JSON formats are all
compatible enough, as far as I'm concerned. There do seem to be some
extra fields in Docker's container configs that aren't present in the
OCI spec; we'll see how big of a problem that is, I guess.