Allow customizing the default tarbuild ModTime
I think it's long past time to start writing some unit tests for
zeroimage, and this value will almost certainly need to be modified in
some future tests. Figured I should get to it while I had it in mind,
and since this is an internal package it's not like I'm *really*
introducing a new API surface.