Fix and test tarbuild path normalization
The original implementation did not actually match my expectations,
since it didn't strip leading ../ components from a cleaned path. The
path normalization process now explicitly interprets all provided paths
relative to the root of the archive, so the final archive will never
look like it's trying to break out of the directory into which it's
extracted.