Set media type when writing manifests
See CVE-2021-41190.

Per the updated OCI distribution spec: "All manifests SHOULD include a
`mediaType` field declaring the type of the manifest being pushed." I
don't see any reason why the manifest in an OCI archive can't also
contain this, so both implementations are going to get it.

The other recommendation is that when pulling a manifest, "clients
SHOULD reject unless the `mediaType` field's value matches the type
specified by the `Content-Type` header." The current setup for loading
images is not well-architected to meet this recommendation, since we
throw away the Content-Type header before invoking the generic loader
and rely entirely on the fields in the manifest to determine its type.
However, the idea of the vulnerability is that a malicious registry can
change the Content-Type between two pulls of the same manifest, get
clients to interpret it differently in different contexts, and convince
the two pulls to reference different images even when the manifest is
pulled by digest (of course this is irrelevant for tags since you can
serve whatever you want whenever you want). So ignoring Content-Type and
only sniffing the manifest should still make zeroimage invulnerable to
the real issue...

...or at least it would if zeroimage actually verified that a manifest
pulled by digest matches the presented digest. Oops. Going to add some
comments about that while I'm thinking about it.