Verify digests of root manifests when known
This should close the hole where a malicious registry could feed
zeroimage a manifest that doesn't match the digest provided on the
command line, and force it to build an image with an arbitrary base
rather than the one intended by the user.

This doesn't apply to OCI archive bases or images pulled by tag, since
those situations don't give us any initial digest to validate against.

The digest verification for the root manifest is centralized in the
generic image loader, just like the digest verifications for nested
manifests and image configs.