Don't detect index if manifest list is literally "null"
This came to mind as I was reviewing the json package documentation. If
this value were null, the json.RawMessage representation would literally
be "null", which is not empty, which could make us think this is an
index even if it isn't. This is admittedly a weird case that would
require a crafted manifest of some sort.