Use containerd package for platform handling
The previous custom handling of platform specifications was not
consistent with the platform values that I'm actually used to from
Docker. With this change, we're now using the same underlying platforms
package from containerd that can intelligently select and order
platforms. This also means I can get rid of the error that occurs when
multiple ARM variants are available, and just make the most sensible
choice the same way that other container tools do.