Note that Docker Desktop has a secure credential store
In the event that Docker is configured to store credentials in a
keychain, e.g. when using Docker for Mac, those credentials will not
actually be stored in the config file itself, so it's confusing to
mention that file in the output. Furthermore, the Docker app needs to be
running in this case to provide access to the credentials, which goes
against the "you don't need Docker" description and is therefore worth
calling out.